+++ src/lib.rs
@@ -66,3 +66,6 @@ pub use numerical::bdf::*;
 
 #[allow(unused_imports)]
 pub use numerical::runge_kutta::*;
+
+#[allow(unused_imports)]
+pub use util::api::*;